+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@psycle/repsycle",
-  "version": "0.0.27-beta.8",
+  "version": "0.0.27-beta.9",
   "description": "Psycle Research front-end toolkit",
   "author": "Psycle Research",
   "keywords": [
