+++ src/draw-zone/index.stories.tsx
@@ -1,4 +1,3 @@
-import { isEmpty } from 'lodash'
 import React, {
     FunctionComponent,
     useCallback,
@@ -7,18 +6,18 @@ import React, {
     useState,
 } from 'react'
 
-import DrawZone, { ChangedElement, DrawZoneContainer, useDrawZone } from '.'
+import DrawZone, { DrawZoneContainer, useControls, useLoadImage } from '.'
+import type { DrawZoneElement } from '.'
+import { DrawZoneShape, PictureLoadingState } from './types'
 
 export default {
     title: 'Components/DrawZone',
     component: DrawZone,
     decorators: [
         (Story: FunctionComponent) => (
-            <DrawZoneContainer>
-                <div style={{ height: '100vh' }}>
-                    <Story />
-                </div>
-            </DrawZoneContainer>
+            <div style={{ height: '100vh' }}>
+                <Story />
+            </div>
         ),
     ],
     parameters: {
@@ -26,833 +25,79 @@ export default {
     },
 }
 
-export function Default() {
-    const originalElements = [
-        {
-            id: 'rect1',
-            points: [
-                { x: 125, y: 94 },
-                { x: 250, y: 1 },
-            ],
-            color: '#00ff00',
-        },
-    ]
+function Controls() {
     const {
-        isMarkerShown,
-        originalSize,
+        contentHidden,
+        markerVisible,
+        move,
+        reset,
+        toggleContent,
+        toggleMarker,
+        toggleMove,
         zoomIn,
         zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
-
-    return (
-        <>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <DrawZone
-                src="https://upload.wikimedia.org/wikipedia/commons/thumb/b/b6/Image_created_with_a_mobile_phone.png/250px-Image_created_with_a_mobile_phone.png"
-                elements={elements}
-                mode={move ? 'move' : 'draw'}
-                shape="rect"
-                onChange={(elements: ChangedElement[]) => setElements(elements)}
-                remove={(id) =>
-                    setElements((el) => el.filter((e) => e.id !== id))
-                }
-            />
-        </>
-    )
-}
-
-export function Rects() {
-    const originalElements = [
-        {
-            id: 'rect1',
-            points: [
-                { x: 125, y: 94 },
-                { x: 250, y: 1 },
-            ],
-            color: '#00ff00',
-        },
-    ]
-    const {
-        isMarkerShown,
-        originalSize,
-        zoomIn,
-        zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
-
+    } = useControls()
     return (
         <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div style={{ backgroundColor: '#aaa' }}>
-                <DrawZone
-                    src="https://upload.wikimedia.org/wikipedia/commons/thumb/b/b6/Image_created_with_a_mobile_phone.png/250px-Image_created_with_a_mobile_phone.png"
-                    elements={elements}
-                    onChange={(elements: ChangedElement[]) =>
-                        setElements(elements)
-                    }
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'draw'}
-                    shape="rect"
-                >
-                    {elements.map((element, index) => {
-                        const elem = element as ChangedElement
-
-                        if (!elem.selected) return null
-
-                        return (
-                            <div
-                                key={index}
-                                style={{
-                                    position: 'absolute',
-                                    zIndex: 10,
-                                    top: elem.rect
-                                        ? `${elem.rect.y}%`
-                                        : `${elem.points[0].y * 100}%`,
-                                    left: elem.rect
-                                        ? `${elem.rect.x}%`
-                                        : `${elem.points[0].x * 100}%`,
-                                    border:
-                                        elem.rect && elem.selected
-                                            ? '1px dashed black'
-                                            : 'none',
-                                    width:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.width}%`
-                                            : 'auto',
-                                    height:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.height}%`
-                                            : 'auto',
-                                    pointerEvents: 'none',
-                                }}
-                            >
-                                <button
-                                    style={{
-                                        pointerEvents: 'auto',
-                                    }}
-                                    onClick={() => {
-                                        setElements(
-                                            elements.filter(
-                                                (_, idx) => index !== idx,
-                                            ),
-                                        )
-                                    }}
-                                >
-                                    supprimer
-                                </button>
-                            </div>
-                        )
-                    })}
-                </DrawZone>
-            </div>
+            <button onClick={reset}>Reset</button>
+            <button onClick={toggleMove}>{move ? 'Move' : 'Draw'}</button>
+            <button onClick={zoomIn}>Zoom in</button>
+            <button onClick={zoomOut}>Zoom out</button>
+            <button onClick={toggleMarker}>
+                {markerVisible ? 'Hide' : 'Show'} marker
+            </button>
+            <button onClick={toggleContent}>
+                {contentHidden ? 'Show' : 'Hide'} content
+            </button>
         </div>
     )
 }
 
-export function Polygons() {
-    const originalElements = [
-        /*
-        {
-            id: 'poly2',
-            points: [
-                { x: 0, y: 0 },
-                { x: 125, y: 125 },
-                { x: 62.5, y: 250 },
-            ],
-        },
-        {
-            id: 'poly3',
-            points: [
-                { x: 25, y: 25 },
-                { x: 100, y: 100 },
-                { x: 62.5, y: 200 },
-            ],
-        },
-        {
-            id: 'poly1',
-            points: [
-                { x: 50, y: 75 },
-                { x: 50, y: 212.5 },
-                { x: 150, y: 200 },
-                { x: 200, y: 50 },
-            ],
-        },
-        {
-            id: 'poly2',
-            points: [
-                { x: 37, y: 26 },
-                { x: 125, y: 188 },
-                { x: 250, y: 188 },
-                { x: 250, y: 94 },
-            ],
-            color: '#00ff00',
-        },
-        */
-    ]
-    const {
-        isMarkerShown,
-        originalSize,
-        zoomIn,
-        zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
-
-    return (
-        <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div style={{ backgroundColor: '#aaa' }}>
-                <DrawZone
-                    src="https://upload.wikimedia.org/wikipedia/commons/thumb/b/b6/Image_created_with_a_mobile_phone.png/250px-Image_created_with_a_mobile_phone.png"
-                    elements={elements}
-                    onChange={(elements: ChangedElement[]) =>
-                        setElements(elements)
-                    }
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'draw'}
-                    shape="poly"
-                >
-                    {elements.map((element, index) => {
-                        const elem = element as ChangedElement
-
-                        if (!elem.selected) return null
-
-                        return (
-                            <div
-                                key={index}
-                                style={{
-                                    position: 'absolute',
-                                    zIndex: 10,
-                                    top: elem.rect
-                                        ? `${elem.rect.y}%`
-                                        : `${elem.points[0].y * 100}%`,
-                                    left: elem.rect
-                                        ? `${elem.rect.x}%`
-                                        : `${elem.points[0].x * 100}%`,
-                                    border:
-                                        elem.rect && elem.selected
-                                            ? '1px dashed black'
-                                            : 'none',
-                                    width:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.width}%`
-                                            : 'auto',
-                                    height:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.height}%`
-                                            : 'auto',
-                                    pointerEvents: 'none',
-                                }}
-                            >
-                                <button
-                                    style={{
-                                        pointerEvents: 'auto',
-                                    }}
-                                    onClick={() => {
-                                        setElements(
-                                            elements.filter(
-                                                (_, idx) => index !== idx,
-                                            ),
-                                        )
-                                    }}
-                                >
-                                    supprimer
-                                </button>
-                            </div>
-                        )
-                    })}
-                </DrawZone>
-            </div>
-        </div>
-    )
+interface StoryZoneElement extends DrawZoneElement {
+    readonly metadata: Record<string, string>
 }
 
-export function ScaleIn() {
-    const originalElements = [
-        {
-            id: 'poly2',
-            points: [
-                { x: 2520, y: 1410 },
-                { x: 1450, y: 3760 },
-                { x: 3550, y: 3580 },
-                { x: 3420, y: 1930 },
-            ],
-            color: '#00ff00',
-        },
-    ]
-    const {
-        isMarkerShown,
-        originalSize,
-        zoomIn,
-        zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
+type EditorProps = {
+    readonly initialElements: StoryZoneElement[]
+    readonly shape: DrawZoneShape
+    readonly src: string
+}
+function Editor({ initialElements, shape, src }: EditorProps) {
+    const { status, pictureSize } = useLoadImage(src)
+    const [elements, setElements] = useState<Array<StoryZoneElement>>([])
 
     useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
-
-    return (
-        <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div
-                style={{
-                    backgroundColor: '#aaa',
-                    width: '500px',
-                    height: '500px',
-                }}
-            >
-                <DrawZone
-                    src="https://picsum.photos/seed/drawzone/2000/1000"
-                    elements={elements}
-                    onChange={(elements: ChangedElement[]) =>
-                        setElements(elements)
-                    }
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'draw'}
-                    shape="poly"
-                    sizeMode="fit"
-                >
-                    {elements.map((element, index) => {
-                        const elem = element as ChangedElement
-
-                        if (!elem.selected) return null
-
-                        return (
-                            <div
-                                key={index}
-                                style={{
-                                    position: 'absolute',
-                                    zIndex: 10,
-                                    top: elem.rect
-                                        ? `${elem.rect.y}%`
-                                        : `${elem.points[0].y * 100}%`,
-                                    left: elem.rect
-                                        ? `${elem.rect.x}%`
-                                        : `${elem.points[0].x * 100}%`,
-                                    border:
-                                        elem.rect && elem.selected
-                                            ? '1px dashed black'
-                                            : 'none',
-                                    width:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.width}%`
-                                            : 'auto',
-                                    height:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.height}%`
-                                            : 'auto',
-                                    pointerEvents: 'none',
-                                }}
-                            >
-                                <button
-                                    style={{
-                                        pointerEvents: 'auto',
-                                    }}
-                                    onClick={() => {
-                                        setElements(
-                                            elements.filter(
-                                                (_, idx) => index !== idx,
-                                            ),
-                                        )
-                                    }}
-                                >
-                                    supprimer
-                                </button>
-                            </div>
-                        )
-                    })}
-                </DrawZone>
-            </div>
-        </div>
-    )
-}
-
-export function ScaleOut() {
-    const originalElements = [
-        {
-            id: 'rect1',
-            points: [
-                { x: 125, y: 94 },
-                { x: 250, y: 1 },
-            ],
-            color: '#00ff00',
-        },
-    ]
-    const {
-        isMarkerShown,
-        originalSize,
-        zoomIn,
-        zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
-
-    return (
-        <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div
-                style={{
-                    backgroundColor: '#aaa',
-                    width: '500px',
-                    height: '500px',
-                }}
-            >
-                <DrawZone
-                    src="https://upload.wikimedia.org/wikipedia/commons/thumb/b/b6/Image_created_with_a_mobile_phone.png/250px-Image_created_with_a_mobile_phone.png"
-                    elements={elements}
-                    onChange={(elements: ChangedElement[]) =>
-                        setElements(elements)
-                    }
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'draw'}
-                    shape="rect"
-                    sizeMode="fit"
-                >
-                    {elements.map((element, index) => {
-                        const elem = element as ChangedElement
-
-                        if (!elem.selected) return null
-
-                        return (
-                            <div
-                                key={index}
-                                style={{
-                                    position: 'absolute',
-                                    zIndex: 10,
-                                    top: elem.rect
-                                        ? `${elem.rect.y}%`
-                                        : `${elem.points[0].y * 100}%`,
-                                    left: elem.rect
-                                        ? `${elem.rect.x}%`
-                                        : `${elem.points[0].x * 100}%`,
-                                    border:
-                                        elem.rect && elem.selected
-                                            ? '1px dashed black'
-                                            : 'none',
-                                    width:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.width}%`
-                                            : 'auto',
-                                    height:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.height}%`
-                                            : 'auto',
-                                    pointerEvents: 'none',
-                                }}
-                            >
-                                <button
-                                    style={{
-                                        pointerEvents: 'auto',
-                                    }}
-                                    onClick={() => {
-                                        setElements(
-                                            elements.filter(
-                                                (_, idx) => index !== idx,
-                                            ),
-                                        )
-                                    }}
-                                >
-                                    supprimer
-                                </button>
-                            </div>
-                        )
-                    })}
-                </DrawZone>
-            </div>
-        </div>
-    )
-}
-
-export function ScaleInRect() {
-    const originalElements = [
-        {
-            id: 'rect1',
-            points: [
-                { x: 125, y: 94 },
-                { x: 250, y: 1 },
-            ],
-            color: '#00ff00',
-        },
-    ]
-    const {
-        isMarkerShown,
-        originalSize,
-        zoomIn,
-        zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
-
-    return (
-        <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div
-                style={{
-                    backgroundColor: '#aaa',
-                    width: '500px',
-                    height: '500px',
-                }}
-            >
-                <DrawZone
-                    src="https://picsum.photos/seed/drawzone/2000/1000"
-                    elements={elements}
-                    onChange={(elements: ChangedElement[]) =>
-                        setElements(elements)
-                    }
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'draw'}
-                    shape="rect"
-                    sizeMode="fit"
-                >
-                    {elements.map((element, index) => {
-                        const elem = element as ChangedElement
-
-                        if (!elem.selected) return null
-
-                        return (
-                            <div
-                                key={index}
-                                style={{
-                                    position: 'absolute',
-                                    zIndex: 10,
-                                    top: elem.rect
-                                        ? `${elem.rect.y}%`
-                                        : `${elem.points[0].y * 100}%`,
-                                    left: elem.rect
-                                        ? `${elem.rect.x}%`
-                                        : `${elem.points[0].x * 100}%`,
-                                    border:
-                                        elem.rect && elem.selected
-                                            ? '1px dashed black'
-                                            : 'none',
-                                    width:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.width}%`
-                                            : 'auto',
-                                    height:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.height}%`
-                                            : 'auto',
-                                    pointerEvents: 'none',
-                                }}
-                            >
-                                <button
-                                    style={{
-                                        pointerEvents: 'auto',
-                                    }}
-                                    onClick={() => {
-                                        setElements(
-                                            elements.filter(
-                                                (_, idx) => index !== idx,
-                                            ),
-                                        )
-                                    }}
-                                >
-                                    supprimer
-                                </button>
-                            </div>
-                        )
-                    })}
-                </DrawZone>
-            </div>
-        </div>
-    )
-}
-
-export function None() {
-    const { isMarkerShown, zoomIn, zoomOut, toggleMarker, reset } =
-        useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    return (
-        <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div
-                style={{
-                    backgroundColor: '#aaa',
-                    width: '500px',
-                    height: '500px',
-                }}
-            >
-                <DrawZone
-                    src="https://picsum.photos/seed/drawzone/2000/1000"
-                    elements={elements}
-                    onChange={(elements: ChangedElement[]) =>
-                        setElements(elements)
-                    }
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'none'}
-                    shape="rect"
-                    sizeMode="fit"
-                />
-            </div>
-        </div>
-    )
-}
-
-export function BugRepro() {
-    const originalElements = [
-        {
-            id: 'rect1',
-            points: [
-                { x: 125, y: 94 },
-                { x: 250, y: 1 },
-            ],
-            color: '#00ff00',
-        },
-    ]
-    const {
-        isMarkerShown,
-        originalSize,
-        zoomIn,
-        zoomOut,
-        toggleMarker,
-        reset,
-    } = useDrawZone()
-    const [move, setMove] = useState(false)
-    const [elements, setElements] = useState<Array<Partial<ChangedElement>>>([])
-
-    useEffect(() => {
-        if (originalSize) {
-            setElements(
-                originalElements.map(
-                    (element) =>
-                        ({
-                            ...element,
-                            points: element.points.map((point) => ({
-                                x: point.x / originalSize.width,
-                                y: point.y / originalSize.height,
-                            })),
-                        } as ChangedElement),
-                ),
-            )
-        }
-    }, [originalSize])
+        setElements(initialElements)
+    }, [initialElements])
 
     const buildMetas = useCallback(
-        (elem: ChangedElement) => {
+        (elem: DrawZoneElement) => {
+            if (!pictureSize || shape !== 'rect') return {}
+
             return {
                 width: Math.floor(
-                    (elem.rect.width * originalSize.width) / 100,
+                    (elem.rect.width * pictureSize.width) / 100,
                 ).toString(),
                 height: Math.floor(
-                    (elem.rect.height * originalSize.height) / 100,
+                    (elem.rect.height * pictureSize.height) / 100,
                 ).toString(),
                 diagonal: Math.floor(
                     Math.hypot(
-                        (elem.rect.height * originalSize.height) / 100,
-                        (elem.rect.width * originalSize.width) / 100,
+                        (elem.rect.height * pictureSize.height) / 100,
+                        (elem.rect.width * pictureSize.width) / 100,
                     ),
                 ).toString(),
                 area: Math.floor(
-                    ((elem.rect.width * originalSize.width) / 100) *
-                        ((elem.rect.height * originalSize.height) / 100),
+                    ((elem.rect.width * pictureSize.width) / 100) *
+                        ((elem.rect.height * pictureSize.height) / 100),
                 ).toString(),
-            }
+            } as Record<string, string>
         },
-        [originalSize],
+        [pictureSize, shape],
     )
 
     const onChange = useCallback(
-        (elements: ChangedElement[]) => {
+        (elements: DrawZoneElement[]) => {
             const newElements = elements.map((elem) => {
                 return {
                     ...elem,
@@ -871,23 +116,36 @@ export function BugRepro() {
     )
 
     return (
-        <div>
-            <div>
-                <button onClick={() => setMove((m) => !m)}>
-                    {move ? 'Déplacer actif' : 'Déplacer inactif'}
-                </button>
-                <button onClick={() => zoomOut()}>Réduire</button>
-                <button onClick={() => zoomIn()}>Agrandir</button>
-                <button onClick={() => toggleMarker()}>
-                    {isMarkerShown ? 'Cacher' : 'Afficher'} marqueur
-                </button>
-                <button onClick={() => reset()}>Reset (taille/position)</button>
-            </div>
-            <div>
-                {element && !isEmpty(element.metadata) && (
-                    <pre><code>{JSON.stringify(element.metadata)}</code></pre>
+        <>
+            <div
+                style={{
+                    display: 'flex',
+                    flexDirection: 'row',
+                    alignItems: 'center',
+                    height: '40px',
+                }}
+            >
+                <span>Current item :</span>
+                {element && (
+                    <>
+                        <button
+                            onClick={() => {
+                                onChange(
+                                    elements.filter(
+                                        ({ id }) => id !== element.id,
+                                    ),
+                                )
+                            }}
+                        >
+                            Delete
+                        </button>
+                        <pre>
+                            <code>{JSON.stringify(element.metadata)}</code>
+                        </pre>
+                    </>
                 )}
             </div>
+
             <div
                 style={{
                     backgroundColor: '#aaa',
@@ -895,68 +153,64 @@ export function BugRepro() {
                     height: '500px',
                 }}
             >
-                <DrawZone
-                    src="https://picsum.photos/seed/drawzone/2000/1000"
-                    elements={elements}
-                    onChange={onChange}
-                    remove={(id) =>
-                        setElements((el) => el.filter((e) => e.id !== id))
-                    }
-                    mode={move ? 'move' : 'draw'}
-                    shape="rect"
-                    sizeMode="fit"
-                >
-                    {elements.map((element, index) => {
-                        const elem = element as ChangedElement
-
-                        if (!elem.selected) return null
-
-                        return (
-                            <div
-                                key={index}
-                                style={{
-                                    position: 'absolute',
-                                    zIndex: 10,
-                                    top: elem.rect
-                                        ? `${elem.rect.y}%`
-                                        : `${elem.points[0].y * 100}%`,
-                                    left: elem.rect
-                                        ? `${elem.rect.x}%`
-                                        : `${elem.points[0].x * 100}%`,
-                                    border:
-                                        elem.rect && elem.selected
-                                            ? '1px dashed black'
-                                            : 'none',
-                                    width:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.width}%`
-                                            : 'auto',
-                                    height:
-                                        elem.rect && elem.selected
-                                            ? `${elem.rect.height}%`
-                                            : 'auto',
-                                    pointerEvents: 'none',
-                                }}
-                            >
-                                <button
-                                    style={{
-                                        pointerEvents: 'auto',
-                                    }}
-                                    onClick={() => {
-                                        setElements(
-                                            elements.filter(
-                                                (_, idx) => index !== idx,
-                                            ),
-                                        )
-                                    }}
-                                >
-                                    supprimer
-                                </button>
-                            </div>
-                        )
-                    })}
-                </DrawZone>
+                {status === PictureLoadingState.Done && (
+                    <DrawZone
+                        src={src}
+                        elements={elements}
+                        fitMode="fit"
+                        mode="draw"
+                        onChange={onChange}
+                        shape={shape}
+                    />
+                )}
             </div>
-        </div>
+        </>
+    )
+}
+
+export function Default() {
+    return (
+        <DrawZoneContainer>
+            <Controls />
+            <Editor
+                src="https://picsum.photos/seed/drawzone/2000/1000"
+                shape="rect"
+                initialElements={[
+                    {
+                        id: 'rect1',
+                        points: [
+                            { x: 125, y: 94 },
+                            { x: 250, y: 1 },
+                        ],
+                        color: [0, 0, 255],
+                        selected: false,
+                    } as unknown as StoryZoneElement,
+                ]}
+            />
+        </DrawZoneContainer>
+    )
+}
+
+export function Poly() {
+    return (
+        <DrawZoneContainer>
+            <Controls />
+            <Editor
+                src="https://picsum.photos/seed/drawzone/2000/1000"
+                shape="poly"
+                initialElements={[
+                    {
+                        id: 'poly2',
+                        points: [
+                            { x: 37, y: 26 },
+                            { x: 125, y: 188 },
+                            { x: 250, y: 188 },
+                            { x: 250, y: 94 },
+                        ],
+                        color: '#00ff00',
+                    } as unknown as StoryZoneElement,
+                ]}
+            />
+        </DrawZoneContainer>
     )
 }
